pyproject.toml
[project]
name = "djlib"
version = "0.0.1"
description = "DJ library management"
readme = "README.md"
authors = [
    { name = "Christian Vanderwall", email = "christian@cvdub.net" }
]
requires-python = ">=3.13"
dependencies = [
    "platformdirs>=4.3.6",
    "requests>=2.32.3",
    "tortoise-orm>=0.24.2",
]

[project.scripts]
djlib = "djlib:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "gnureadline>=8.2.13",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
